requirements.txt
langchain==0.1.12
python-docx==1.1.0
PyPDF2==3.0.1
chromadb==0.4.24
sentence-transformers==2.5.1
python-dotenv==1.0.1
openai==1.14.0
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9 